doc: Give @mdash macro an argument
This macro is used to consume whitespace so that the following sentence can
start on the next line without producing any whitespace in the output.  Its
argument is, therefore, whitespace.

This used to work in earlier versions of Texinfo, but around 6.{6,7} it
began failing because an argument was provided when it wasn't defined with
one.